+++ src/commands.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { updateContextKeys } from './contextTracker';
+import { updateContextKeys } from './contextKeys';
 import { RunGroup } from './enums';
 import { log } from './util/logging';
